Makefile, stockClient.go, targetAnnualizedReturn.go, currentAnnualizedReturn.go: added makefile and broke out functionality of current+target annualized returns from stockClient
